refactor: split pending summary store contracts
